Add property task.startImmediately
Starts indexing tasks even if there are no incoming events. This is
useful if you have a data source that doesn't have a lot of events, but
you want to be ready when the events start coming in so that they can be
ingested quickly without waiting for indexing task to start.